invariants of player claiming

+++ pointercrate-demonlist/src/player/claim/put.rs
@@ -1,5 +1,5 @@
 use crate::{
-    error::Result,
+    error::{DemonlistError, Result},
     player::{claim::PlayerClaim, DatabasePlayer},
 };
 use log::info;
@@ -7,6 +7,19 @@ use sqlx::PgConnection;
 
 impl DatabasePlayer {
     pub async fn initiate_claim(&self, claimed_by: i32, connection: &mut PgConnection) -> Result<PlayerClaim> {
+        // check if player is already claimed and verified
+        let is_claimed = sqlx::query!(
+            r#"SELECT EXISTS(SELECT 1 FROM player_claims WHERE player_id = $1 AND verified) AS "is_claimed!""#,
+            self.id
+        )
+        .fetch_one(&mut *connection)
+        .await?
+        .is_claimed;
+
+        if is_claimed {
+            return Err(DemonlistError::AlreadyClaimed)
+        }
+
         // first, clear all claims by the given user
         let result = sqlx::query!("DELETE FROM player_claims WHERE member_id = $1", claimed_by)
             .execute(&mut *connection)
